backend/requirements.txt
Flask
flask_jwt_extended
flask_sqlalchemy
pandas
Werkzeug
flask_mail
flask-cors
watchdog
python-dotenv
openai